requirements.txt
django-easy-audit
asgiref==3.2.4
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
django==3.2.6
django-crispy-forms==1.12.0
django-embed-video==1.4.0
django-filter==2.4.0
django-freezer==0.1.4
django-grappelli==2.15.1
django-import-export==2.5.0
django-js-routes==0.1.4
django-password-validators==1.6.1
django-seed==0.2.2
django-session-timeout==0.1.0
django-suit==0.2.28
Django-Verify-Email==1.0.7
djangorestframework==3.12.4
et-xmlfile==1.1.0
Faker==8.9.1
idna==2.10
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
django-widget-tweaks==1.4.8
django-xhtml2pdf